+++ KamlaGuti/Assets/ML-Agents/Timers/Game_timers.json
@@ -1,1 +1,1 @@
-{"count":1,"self":31.7137568,"total":32.1548393,"children":{"AgentSendState":{"count":27372,"self":0.1830299,"total":0.2600437,"children":{"CollectObservations":{"count":11390,"self":0.0620133,"total":0.0620133,"children":null},"CollectDiscreteActionMasks":{"count":11390,"self":0.0150005,"total":0.0150005,"children":null}}},"DecideAction":{"count":27372,"self":0.060012699999999995,"total":0.060012699999999995,"children":null},"AgentAct":{"count":27372,"self":0.1200262,"total":0.1200262,"children":null}},"gauges":{"GutiBehaviour.CumulativeReward":{"count":31,"max":-9.4375,"min":-69.0625,"runningAverage":-26.4213715,"value":-22.9375,"weightedAverage":-22.5572453}},"metadata":{"timer_format_version":"0.1.0","start_time_seconds":"1612690897","unity_version":"2020.1.3f1","command_line_arguments":"C:\\Program Files\\Unity\\Hub\\Editor\\2020.1.3f1\\Editor\\Unity.exe -projectpath C:\/Unity\/KamalaGuit\/KamalaGuit\/KamlaGuti -useHub -hubIPC -cloudEnvironment production -hubSessionId 93abfec0-62d9-11eb-8353-f1374879b3e6 -accessToken 3RiypPlqDcrgrKobSe-LVjgspp-xHwGq_Bo-rBhsWwk011f","communication_protocol_version":"1.0.0","com.unity.ml-agents_version":"1.0.5","scene_name":"Game","end_time_seconds":"1612690929"}}
+{"count":1,"self":35.913647999999995,"total":36.3780149,"children":{"AgentSendState":{"count":31607,"self":0.19704919999999998,"total":0.27334179999999997,"children":{"CollectObservations":{"count":14279,"self":0.0542871,"total":0.0542871,"children":null},"CollectDiscreteActionMasks":{"count":14279,"self":0.022005499999999997,"total":0.022005499999999997,"children":null}}},"DecideAction":{"count":31607,"self":0.057010399999999996,"total":0.057010399999999996,"children":null},"AgentAct":{"count":31607,"self":0.13401559999999998,"total":0.13401559999999998,"children":null}},"gauges":{"GutiBehaviour.CumulativeReward":{"count":40,"max":-5.6875,"min":-70.875,"runningAverage":-24.9312515,"value":-49.625,"weightedAverage":-30.2737617}},"metadata":{"timer_format_version":"0.1.0","start_time_seconds":"1612728790","unity_version":"2020.1.3f1","command_line_arguments":"C:\\Program Files\\Unity\\Hub\\Editor\\2020.1.3f1\\Editor\\Unity.exe -projectpath C:\/Unity\/KamalaGuit\/KamalaGuit\/KamlaGuti -useHub -hubIPC -cloudEnvironment production -hubSessionId 93abfec0-62d9-11eb-8353-f1374879b3e6 -accessToken 3RiypPlqDcrgrKobSe-LVjgspp-xHwGq_Bo-rBhsWwk011f","communication_protocol_version":"1.0.0","com.unity.ml-agents_version":"1.0.5","scene_name":"Game","end_time_seconds":"1612728826"}}
