+++ KamlaGuti/Assets/Scripts/AI/MinMaxAI.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using Board.Guti;
 using Random = UnityEngine.Random;
 
 public class MinMaxAi
