+++ KamlaGuti/Assets/Scripts/Board/View/BoardGui.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: b089ae01dd76461e99caccbf16671f7c
+timeCreated: 1612728174
