+++ KamlaGuti/Assets/Scripts/Board/View/Guti.cs
@@ -1,53 +1,57 @@
-﻿using UnityEngine;
+﻿using Board.Guti;
+using UnityEngine;
 
-public class Guti : MonoBehaviour
+namespace Board.View
 {
-    internal Address address;
-    internal GutiType gutiType = GutiType.NoGuti;
-
-    public void SetAddress(Address address, float scale = 2.0f)
+    public class Guti : MonoBehaviour
     {
-        this.address = address;
-        var o = gameObject;
-        o.transform.position= new Vector3(address.x, address.y, -1);
-        o.transform.localScale = new Vector3(scale, scale, 1);
-    }
+        internal Address address;
+        internal GutiType gutiType = GutiType.NoGuti;
 
-    public void SetScale(float scale)
-    {
-        gameObject.transform.localScale = new Vector3(scale, scale, 1);
-    }
-
-    public void SetGutiType(GutiType gutiType)
-    {
-        this.gutiType = gutiType;
-        var spriteRenderer = gameObject.GetComponent<SpriteRenderer>();
-        switch (gutiType)
+        public void SetAddress(Address address, float scale = 2.0f)
         {
-            case GutiType.RedGuti:
-                spriteRenderer.color = Color.red;
-                gameObject.name = $"Red-x{address.x}-y{address.y}";
-                break;
-            case GutiType.GreenGuti:
-                gameObject.name = $"Green-x{address.x}-y{address.y}";
-                spriteRenderer.color = Color.green;
-                break;
-            case GutiType.NoGuti:
-                Destroy(gameObject);
-                break;
-            case  GutiType.Highlight:
-                spriteRenderer.color = Color.yellow;
-                break;
-            default:
-                Destroy(gameObject);
-                break;
+            this.address = address;
+            var o = gameObject;
+            o.transform.position= new Vector3(address.x, address.y, -1);
+            o.transform.localScale = new Vector3(scale, scale, 1);
         }
-    }
 
-    public void SetGutiColor(Color color)
-    {
-        var spriteRenderer = gameObject.GetComponent<SpriteRenderer>();
-        spriteRenderer.color = color;
-    }
+        public void SetScale(float scale)
+        {
+            gameObject.transform.localScale = new Vector3(scale, scale, 1);
+        }
 
+        public void SetGutiType(GutiType gutiType)
+        {
+            this.gutiType = gutiType;
+            var spriteRenderer = gameObject.GetComponent<SpriteRenderer>();
+            switch (gutiType)
+            {
+                case GutiType.RedGuti:
+                    spriteRenderer.color = Color.red;
+                    gameObject.name = $"Red-x{address.x}-y{address.y}";
+                    break;
+                case GutiType.GreenGuti:
+                    gameObject.name = $"Green-x{address.x}-y{address.y}";
+                    spriteRenderer.color = Color.green;
+                    break;
+                case GutiType.NoGuti:
+                    Destroy(gameObject);
+                    break;
+                case  GutiType.Highlight:
+                    spriteRenderer.color = Color.yellow;
+                    break;
+                default:
+                    Destroy(gameObject);
+                    break;
+            }
+        }
+
+        public void SetGutiColor(Color color)
+        {
+            var spriteRenderer = gameObject.GetComponent<SpriteRenderer>();
+            spriteRenderer.color = color;
+        }
+
+    }
 }
